api/requirements.txt
Django==2.1.3
djangorestframework==3.9.0
djoser==1.3.3
Keras==2.2.4
mysqlclient==1.3.13
Pillow==5.3.0
redis==2.10.6
requests==2.20.1
tensorflow==1.12.0